The new importer modes are now merged into the Plant Portal importer. New modes are - New mode allows user to prevent any rows from uploading if there are any errors at all - New functionality to assist the user to use the Sample External Key to identify samples when uploading occurrences.
Code looks to be working, but could do with some cleanup to simplify

+++ helper_base.php
@@ -1315,7 +1315,7 @@ class helper_base extends helper_config {
    * @param string $service Path to the service URL used. Default is data/handle_media, but could be import/upload_csv.
    * @return string Error message, or true if successful.
    */
-  protected static function send_file_to_warehouse($path, $persist_auth=false, $readAuth = null, $service='data/handle_media') {
+  protected static function send_file_to_warehouse($path, $persist_auth=false, $readAuth = null, $service='data/handle_media',$removeLocalCopy=true) {
     if ($readAuth==null) $readAuth=$_POST;
     $interim_image_folder = isset(parent::$interim_image_folder) ? parent::$interim_image_folder : 'upload/';
     $interim_path = dirname(__FILE__).'/'.$interim_image_folder;
@@ -1346,7 +1346,9 @@ class helper_base extends helper_config {
       }
     }
     //remove local copy
-    unlink(realpath($interim_path.$path));
+    if ($removeLocalCopy==true) {
+      unlink(realpath($interim_path.$path));
+    }
     return $r;
   }
